Cleaned up types in the preprocessor
