+++ Rubberduck.Parsing/Preprocessing/ConditionalCompilationConstantExpression.cs
@@ -1,21 +1,27 @@
-﻿namespace Rubberduck.Parsing.PreProcessing
+﻿using Antlr4.Runtime;
+using System.Collections.Generic;
+
+namespace Rubberduck.Parsing.PreProcessing
 {
     public sealed class ConditionalCompilationConstantExpression : Expression
     {
         private readonly IExpression _tokenText;
         private readonly IExpression _identifier;
         private readonly IExpression _expression;
+        private readonly IExpression _tokens;
         private readonly SymbolTable<string, IValue> _symbolTable;
 
         public ConditionalCompilationConstantExpression(
             IExpression tokenText,
             IExpression identifier, 
             IExpression expression,
+            IExpression tokens,
             SymbolTable<string, IValue> symbolTable)
         {
             _tokenText = tokenText;
             _identifier = identifier;
             _expression = expression;
+            _tokens = tokens;
             _symbolTable = symbolTable;
         }
 
@@ -27,7 +33,8 @@
             _symbolTable.Add(identifier, constantValue);
             return new LivelinessExpression(
                 isAlive: new ConstantExpression(new BoolValue(false)),
-                code: _tokenText).Evaluate();
+                code: _tokenText,
+                tokens: _tokens).Evaluate();
         }
     }
 }
